SI-4552 feat(ui): DI в MyApp
