+++ OOP Patterns/aquarium/lib/logger/res/log_res.dart
@@ -1,14 +1,15 @@
 import 'package:aquarium/pool/pool_state.dart';
+import 'package:aquarium/res/strings.dart';
 
 /// Строки для логирования
 class LogRes {
   static const staffCleanPool = 'Персонал почистил бассейн';
   static const staffServeFishes = 'Персонал обслужил рыб';
   static const staffSetNormalTemp =
-      'Персонал установил нормальную температуру $normalTemperature C';
+      'Персонал установил нормальную температуру $normalTemperature ${AppStrings.celsiusDegree}';
 
   static String changingTemp(double temp) =>
-      'Температура в бассейне стала $temp C';
+      'Температура в бассейне стала $temp ${AppStrings.celsiusDegree}';
   static const fishBirth = 'Родилась новая рыба';
 
   static const oldFishDeath = 'Рыба умерла от старости';
